+++ src/Symfony/Component/Notifier/Transport/AbstractTransportFactory.php
@@ -11,6 +11,7 @@
 
 namespace Symfony\Component\Notifier\Transport;
 
+use Symfony\Component\EventDispatcher\Event;
 use Symfony\Component\EventDispatcher\LegacyEventDispatcherProxy;
 use Symfony\Component\Notifier\Exception\IncompleteDsnException;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
@@ -18,7 +19,6 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 /**
  * @author Konstantin Myakshin <molodchick@gmail.com>
- *
  * @author Fabien Potencier <fabien@symfony.com>
  *
  * @experimental in 5.0
@@ -30,7 +30,7 @@ abstract class AbstractTransportFactory implements TransportFactoryInterface
 
     public function __construct(EventDispatcherInterface $dispatcher = null, HttpClientInterface $client = null)
     {
-        $this->dispatcher = LegacyEventDispatcherProxy::decorate($dispatcher);
+        $this->dispatcher = class_exists(Event::class) ? LegacyEventDispatcherProxy::decorate($dispatcher) : $dispatcher;
         $this->client = $client;
     }