+++ src/Symfony/Component/EventDispatcher/CHANGELOG.md
@@ -1,6 +1,11 @@
 CHANGELOG
 =========
 
+5.1.0
+-----
+
+ * The `LegacyEventDispatcherProxy` class has been deprecated.
+
 5.0.0
 -----